refactor: split source into types/retry/index and enforce 100% coverage
- Split src/index.ts into src/types.ts (public types), src/retry.ts (the
  retry function, pure helpers and the recursive runner) and src/index.ts
  (entry point re-exporting retry plus the types).
- Separate the public retry() from an internal run() loop so the recursion
  carries ResolvedOptions, removing an unreachable nullable branch.
- Add tests for the abort-with-no-reason fallback and listener cleanup; raise
  the coverage thresholds to 100% (statements, branches, functions, lines).
- README: add a "Working with Claude Code" section; update CLAUDE.md to
  describe the new file layout.
